+++ app/src/main/java/com/intempt/core/types/SealedTypes.kt
@@ -128,8 +128,6 @@ internal sealed class EventType(val value: String) {
 
     data object Record : EventType("record")
 
-    data object Alias : EventType("alias")
-
     data object Consent : EventType("consent")
 
     data object Product : EventType("product")